Runtime: fixed possible NPE

+++ org.qiweb/org.qiweb.runtime/src/main/java/org/qiweb/runtime/server/HttpOnDiskRequestAggregator.java
@@ -221,6 +221,9 @@ public class HttpOnDiskRequestAggregator
     public void channelInactive( ChannelHandlerContext ctx )
         throws IOException
     {
-        Files.delete( bodyFile );
+        if( bodyFile != null )
+        {
+            Files.delete( bodyFile );
+        }
     }
 }
